+++ src/net-util/tests/StatusResponse.test.js
@@ -1,7 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { FullResponse, HttpHeaders, IncomingRequest, OriginAddress,
+import { FullResponse, HttpHeaders, IncomingRequest, EndpointAddress,
   RequestContext, StatusResponse }
   from '@this/net-util';
 
@@ -16,7 +16,7 @@ import { FullResponse, HttpHeaders, IncomingRequest, OriginAddress,
 function makeRequest(authority, path) {
   const context = new RequestContext(
     Object.freeze({}),
-    new OriginAddress('10.0.0.1', 10321));
+    new EndpointAddress('10.0.0.1', 10321));
 
   return new IncomingRequest({
     context,